wip: migrate to swagger-generated DTOs, restore read-only PR review UI, bump to IDEA 2026.2.1 / Gitea 1.27.1+
Continues the in-progress DTO migration left mid-flight in the working tree:
replaces the hand-rolled api/rest/models DTO layer with the swagger-v2-spec-
generated api/rest/dto types end to end, fixes the resulting compile breaks
(missing fromDto factories, date-type mismatches, a generator syntax defect
in MergePullRequestOption), and restores + adapts the previously-deleted
read-only PR review UI (list/filters, details tab, diff viewer with inline
comment display) on top of it. Review/comment submission (Milestone 2) is
intentionally left out for now.

Also raises platform floors: pluginSinceBuild 261->262 / platformVersion
2026.2.1 (bounded to 262.* since collaboration-tools APIs are unstable
across platform versions), and Gitea earliestSupportedVersion 1.27.0->1.27.1
with actual enforcement wired into the login flow (previously unenforced).
Needed extra fixes discovered along the way: declare newly-required
platform modules (vcs.dvcs.impl, vcs.impl, their .shared variants) that
2026.2 stopped exposing implicitly, and bump the IntelliJ Platform Gradle
Plugin 2.13.1->2.18.1 since the old version can't parse 2026.2.1's module
descriptor format.

Known issue, not yet fixed (see PR_REVIEW_MIGRATION_PLAN.md): MyPluginTest
(unrelated template leftover) fails on a Jackson version conflict from the
newly-added jackson-datatype-jsr310 dependency; all 51 real Gitea tests pass.

+++ src/main/kotlin/com/github/jpmand/idea/plugin/gitea/pullrequest/diff/GiteaPRDiffViewModel.kt
@@ -37,7 +37,7 @@ class GiteaPRDiffViewModel(
         cs.launch(Dispatchers.IO) {
             _changesState.value = ComputedResult.loading()
             try {
-                val files = repository.loadChangedFiles(pr.number)
+                val files = repository.loadChangedFiles(pr.number.toInt())
                 val fileVms = files.map { file ->
                     GiteaPRDiffFileViewModel(cs, project, repository, file, pr.base.sha, pr.head.sha)
                 }
